requirements.txt
Django==4.2.0
psycopg2-binary==2.9.6
python-decouple==3.8
whitenoise==6.4.0